+++ src/Symfony/Component/Validator/Tests/Fixtures/BackedEnumA.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Symfony\Component\Validator\Tests\Fixtures;
+
+enum BackedEnumA: string
+{
+    case READ = 'read';
+}